requirements.txt
# Synthetic ITSM Incident Data Engine v2.0 - Requirements
# ================================================================
# Production-ready dependencies for Azure OpenAI-powered
# synthetic ServiceNow incident data generation
#
# Author: Hans Havlik
# Organization: Capgemini
# Application: GADM Work Assistant v1.1.4
# Date: December 2024
# ================================================================

# Core Azure OpenAI and LangChain Dependencies
# ---------------------------------------------
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.38
openai>=1.12.0

# Azure Services Integration
# ---------------------------
azure-core>=1.29.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-monitor-opentelemetry>=1.2.0

# Data Processing and Export
# ---------------------------
pandas>=2.1.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
numpy>=1.24.0

# Environment and Configuration Management
# -----------------------------------------
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP and API Clients
# ---------------------
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Async Programming and Concurrency
# ----------------------------------
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# System Monitoring and Performance
# ----------------------------------
psutil>=5.9.0
memory-profiler>=0.61.0

# Utilities and Data Validation
# ------------------------------
jsonschema>=4.20.0
marshmallow>=3.20.0
faker>=20.1.0
uuid>=1.30

# Logging and Monitoring
# ----------------------
structlog>=23.2.0
coloredlogs>=15.0.1
rich>=13.7.0

# Development and Testing (Optional)
# -----------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
coverage>=7.3.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Type Checking and Code Quality
# -------------------------------
typing-extensions>=4.8.0
types-requests>=2.31.0
types-PyYAML>=6.0.0

# Additional Utilities
# --------------------
click>=8.1.0
tqdm>=4.66.0
pathlib2>=2.3.7
dataclasses-json>=0.6.0

# Security and Encryption
# ------------------------
cryptography>=41.0.0
keyring>=24.3.0

# Production Deployment
# ---------------------
gunicorn>=21.2.0
uvicorn>=0.24.0
fastapi>=0.104.0

# Optional: Jupyter Notebook Support
# -----------------------------------
jupyter>=1.0.0
ipykernel>=6.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Version Compatibility Notes:
# ============================
# - Python 3.8+ required
# - Azure OpenAI API version: 2024-02-15-preview or newer
# - LangChain Framework: Latest stable version recommended
# - All packages tested with Python 3.11
#
# Installation Commands:
# ======================
# pip install -r requirements.txt
# pip install -r requirements.txt --upgrade
#
# For development environment:
# pip install -r requirements.txt[dev]
#
# For minimal production:
# pip install langchain langchain-openai pandas openpyxl python-dotenv requests